+++ src/widgets/notepreviewwidget.cpp
@@ -1,0 +1,25 @@
+/*
+ * Copyright (C) 2016 Patrizio Bekerle -- http://www.bekerle.com
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 2 of the License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+ * for more details.
+ */
+
+#include "notepreviewwidget.h"
+
+
+NotePreviewWidget::NotePreviewWidget(QWidget *parent) : QTextBrowser(parent) {
+}
+
+void NotePreviewWidget::resizeEvent(QResizeEvent* event) {
+    emit resize(event->size(), event->oldSize());
+
+    // we need this, otherwise the preview is always blank
+    QTextBrowser::resizeEvent(event);
+}