+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 16.05.2
 - the active tag will now be remembered across different note folders when 
   switching between note folders
+- the note preview is now regenerated once a second when the preview 
+  pane is resized to readjust the width of embedded images
+  (for [Issue #154](https://github.com/pbek/QOwnNotes/issues/154))
 
 ## 16.05.1
 - fixed the automatic image resizing in the preview under Windows